feat(GenericContainer): add PrivilegeSetting trait for managing container privileged mode
