+++ tests/Unit/Containers/GenericContainerTest.php
@@ -123,14 +123,4 @@ class GenericContainerTest extends TestCase
         /** @noinspection PhpUnhandledExceptionInspection */
         $container->start();
     }
-
-    public function testStartWithPrivilegedMode()
-    {
-        $container = (new GenericContainer('alpine:latest'))
-            ->withPrivilegedMode(true);
-        /** @noinspection PhpUnhandledExceptionInspection */
-        $instance = $container->start();
-
-        $this->assertSame(true, $instance->getPrivilegedMode());
-    }
 }
